sample routing examples
