+++ src/Samples/RoutingSample.DirectConsumer/RoutingSample.DirectConsumer.csproj
