+++ src/Samples/RoutingSample.QueueConsumer/Program.cs
@@ -5,22 +5,16 @@ using Twino.Client.TMQ;
 using Twino.Client.TMQ.Connectors;
 using Twino.Protocols.TMQ;
 
-namespace RoutingSample.Consumer
+namespace RoutingSample.QueueConsumer
 {
 	internal class Program
 	{
 		private static void Main(string[] args)
 		{
-			TmqStickyConnector connector = new TmqStickyConnector(TimeSpan.FromSeconds(2), () =>
-			{
-				TmqClient client = new TmqClient();
-				client.SetClientType("SAMPLE-MESSAGE-CONSUMER");
-				return client;
-			});
+			TmqStickyConnector connector = new TmqStickyConnector(TimeSpan.FromSeconds(2));
 			connector.AddHost("tmq://localhost:15500");
 			connector.ContentSerializer = new NewtonsoftContentSerializer();
 			connector.Observer.RegisterConsumer<SampleMessageQueueConsumer>();
-			connector.Observer.RegisterConsumer<SampleDirectMessageConsumer>();
 			connector.Connected += (c) =>
 			{
 				Console.WriteLine("CONNECTED");
@@ -39,15 +33,6 @@ namespace RoutingSample.Consumer
 		public Task Consume(TwinoMessage message, SampleMessage model, TmqClient client)
 		{
 			Console.WriteLine("SAMPLE QUEUE MESSAGE RECEIVED");
-			return Task.CompletedTask;
-		}
-	}
-
-	public class SampleDirectMessageConsumer : IDirectConsumer<SampleMessage>
-	{
-		public Task Consume(TwinoMessage message, SampleMessage model, TmqClient client)
-		{
-			Console.WriteLine("SAMPLE DIRECT MESSAGE RECEIVED");
 			return Task.CompletedTask;
 		}
 	}
