+++ src/Samples/RoutingSample.QueueConsumer/RoutingSample.QueueConsumer.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\..\Twino.Client.TMQ\Twino.Client.TMQ.csproj" />
+      <ProjectReference Include="..\RoutingSample.Models\RoutingSample.Models.csproj" />
+    </ItemGroup>
+
+</Project>
